Add import functionality to the top menu bar
This commit adds the import capability to the top menu bar, while
also removing the need for a menu.js file to load the menu.
