+++ index.html
@@ -10,7 +10,6 @@
     <script src="js/react.js"></script>
     <script src="js/react-dom.js"></script>
     <script src="js/browser.min.js"></script>
-    <script src="js/menubar.js"></script>
     <script src="js/db-init.js"></script>
     <script type="text/babel" src="js/fileupload.js"></script>
     <div id="content"></div>
