+++ js/fileupload.js
@@ -3,7 +3,10 @@
  *               the more traditional click and open file upload system.
  * @author: Ian Hoegen
  ******************************************************************************/
-var Dropzone = require('react-dropzone');
+const Dropzone = require('react-dropzone');
+const app = require('electron');
+const {remote} = require('electron');
+const {Menu} = remote;
 var uploadedFile;
 var uploadedFileContent;
 var FileUploader = React.createClass({
@@ -23,8 +26,9 @@ var FileUploader = React.createClass({
   }
 
 });
-
-ReactDOM.render(<FileUploader />, document.getElementById('content'));
+function renderFileUpload() {
+  ReactDOM.render(<FileUploader />, document.getElementById('content'));
+}
 /** *****************************************************************************
  * @description: This function reads a file and returns the text that is
  *               contained within the file.
@@ -40,3 +44,182 @@ function readFile(file) {
     uploadedFileContent = openedFile;
   };
 }
+/** *****************************************************************************
+ *@author: Ian Hoegen
+ *@description: The JSON outlines a template for the menu, and menu items can
+ *              be added from here.
+ ******************************************************************************/
+const template = [
+  {
+    label: 'File',
+    submenu: [
+      {
+        label: 'Import Project',
+        click(item, focusedWindow) {
+          renderFileUpload();
+        }
+      }
+    ]
+  },
+  {
+    label: 'Edit',
+    submenu: [
+      {
+        label: 'Cut',
+        accelerator: 'CmdOrCtrl+X',
+        role: 'cut'
+      },
+      {
+        label: 'Copy',
+        accelerator: 'CmdOrCtrl+C',
+        role: 'copy'
+      },
+      {
+        label: 'Paste',
+        accelerator: 'CmdOrCtrl+V',
+        role: 'paste'
+      },
+      {
+        label: 'Delete',
+        role: 'delete'
+      },
+      {
+        label: 'Select All',
+        accelerator: 'CmdOrCtrl+A',
+        role: 'selectall'
+      }
+    ]
+  },
+  {
+    label: 'View',
+    submenu: [
+      {
+        label: 'Reload',
+        accelerator: 'CmdOrCtrl+R',
+        click(item, focusedWindow) {
+          if (focusedWindow) focusedWindow.reload();
+        }
+      },
+      {
+        label: 'Toggle Full Screen',
+        accelerator: process.platform === 'darwin' ? 'Ctrl+Command+F' : 'F11',
+        click(item, focusedWindow) {
+          if (focusedWindow)
+            focusedWindow.setFullScreen(!focusedWindow.isFullScreen());
+        }
+      },
+      {
+        label: 'Toggle Developer Tools',
+        accelerator: process.platform === 'darwin' ? 'Alt+Command+I' : 'Ctrl' +
+        '+Shift+I',
+        click(item, focusedWindow) {
+          if (focusedWindow)
+            focusedWindow.webContents.toggleDevTools();
+        }
+      }
+    ]
+  },
+  {
+    label: 'Window',
+    role: 'window',
+    submenu: [
+      {
+        label: 'Minimize',
+        accelerator: 'CmdOrCtrl+M',
+        role: 'minimize'
+      },
+      {
+        label: 'Close',
+        accelerator: 'CmdOrCtrl+W',
+        role: 'close'
+      }
+    ]
+  },
+  {
+    label: 'Help',
+    role: 'help',
+    submenu: [
+      {
+        label: 'Learn More',
+        click() {
+          require('electron').shell.openExternal('https://github.com/WycliffeAssociates/8woc/');
+        }
+      }
+    ]
+  }
+];
+
+if (process.platform === 'darwin') {
+  const name = require('electron').remote.app.getName();
+  template.unshift({
+    label: name,
+    submenu: [
+      {
+        label: 'About ' + name,
+        role: 'about'
+      },
+      {
+        type: 'separator'
+      },
+      {
+        label: 'Services',
+        role: 'services',
+        submenu: []
+      },
+      {
+        type: 'separator'
+      },
+      {
+        label: 'Hide ' + name,
+        accelerator: 'Command+H',
+        role: 'hide'
+      },
+      {
+        label: 'Hide Others',
+        accelerator: 'Command+Alt+H',
+        role: 'hideothers'
+      },
+      {
+        label: 'Show All',
+        role: 'unhide'
+      },
+      {
+        type: 'separator'
+      },
+      {
+        label: 'Quit',
+        accelerator: 'Command+Q',
+        click() {
+          app.quit();
+        }
+      }
+    ]
+  });
+  // Window menu.
+  template[3].submenu = [
+    {
+      label: 'Close',
+      accelerator: 'CmdOrCtrl+W',
+      role: 'close'
+    },
+    {
+      label: 'Minimize',
+      accelerator: 'CmdOrCtrl+M',
+      role: 'minimize'
+    },
+    {
+      label: 'Zoom',
+      role: 'zoom'
+    },
+    {
+      type: 'separator'
+    },
+    {
+      label: 'Bring All to Front',
+      role: 'front'
+    }
+  ];
+}
+
+const menu = Menu.buildFromTemplate(template);
+Menu.setApplicationMenu(menu);
